requirements.txt
python-dotenv
pandas
numpy
pymongo
pymongo[srv]==3.12
certifi
scikit-learn
dill
pyaml

-e .